refactor: add common pack for entities
