+++ ewm/src/main/java/ru/practicum/ewm/category/service/CategoryAdminServiceImpl.java
@@ -3,8 +3,8 @@ package ru.practicum.ewm.category.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.ewm.category.CategoryMapper;
-import ru.practicum.ewm.category.CategoryValidator;
+import ru.practicum.ewm.category.common.CategoryMapper;
+import ru.practicum.ewm.category.common.CategoryValidator;
 import ru.practicum.ewm.category.dto.CategoryDto;
 import ru.practicum.ewm.category.dto.NewCategoryDto;
 import ru.practicum.ewm.category.model.Category;
