+++ ewm/src/main/java/ru/practicum/ewm/category/service/CategoryPublicServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.ewm.category.CategoryMapper;
-import ru.practicum.ewm.category.CategoryValidator;
+import ru.practicum.ewm.category.common.CategoryMapper;
+import ru.practicum.ewm.category.common.CategoryValidator;
 import ru.practicum.ewm.category.dto.CategoryDto;
 import ru.practicum.ewm.category.model.Category;
 import ru.practicum.ewm.category.repository.CategoryRepository;
