+++ ewm/src/main/java/ru/practicum/ewm/compilation/service/CompilationAdminServiceImpl.java
@@ -3,8 +3,8 @@ package ru.practicum.ewm.compilation.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.ewm.compilation.CompilationMapper;
-import ru.practicum.ewm.compilation.CompilationValidator;
+import ru.practicum.ewm.compilation.common.CompilationMapper;
+import ru.practicum.ewm.compilation.common.CompilationValidator;
 import ru.practicum.ewm.compilation.dto.CompilationDto;
 import ru.practicum.ewm.compilation.dto.NewCompilationDto;
 import ru.practicum.ewm.compilation.model.Compilation;
