+++ ewm/src/main/java/ru/practicum/ewm/compilation/service/CompilationPublicServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.ewm.compilation.CompilationMapper;
-import ru.practicum.ewm.compilation.CompilationValidator;
+import ru.practicum.ewm.compilation.common.CompilationMapper;
+import ru.practicum.ewm.compilation.common.CompilationValidator;
 import ru.practicum.ewm.compilation.dto.CompilationDto;
 import ru.practicum.ewm.compilation.model.Compilation;
 import ru.practicum.ewm.compilation.repository.CompilationRepository;
