+++ ewm/src/main/java/ru/practicum/ewm/event/service/EventAdminServiceImpl.java
@@ -5,7 +5,7 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.ewm.category.CategoryMapper;
+import ru.practicum.ewm.category.common.CategoryMapper;
 import ru.practicum.ewm.category.service.CategoryPublicService;
 import ru.practicum.ewm.event.dto.AdminUpdateEventRequest;
 import ru.practicum.ewm.event.common.EventMapper;
